Rename class file
